on est dans la merde

+++ code/ExclusionCheck.h
@@ -12,9 +12,9 @@ public:
 	ExclusionCheck();
 
 private:
-	vector<pair<int, int>> getTDt(vector<int> interTime, vector<int> tabInter, vector<Intervention> interventions);
-	vector<bool> testIfInSeason(vector<int> tInSeasons, vector<pair<int, int>> tDt);
-	vector<int> getConflic(vector<int> exclusion, vector<bool> interIsInSeason, vector<pair<int, int>>  tDts, pair<int, int> tDt, int j);
+	vector<pair<int, double>> getTDt(vector<int> interTime, vector<double> tabInter, vector<Intervention> interventions);
+	vector<bool> testIfInSeason(vector<double> tInSeasons, vector<pair<int, double>> tDt);
+	vector<int> getConflic(vector<double> exclusion, vector<bool> interIsInSeason, vector<pair<int, double>>  tDts, pair<int, double> tDt, int j);
 	dataCollector data;
 
 };
